requirements.txt
flask
flask-cors
dnspython
redis
psycopg2-binary
sqlalchemy
python-dotenv
boto3
requests
openpyxl
python-multipart
